Bugfix for ingest pipeline

+++ cli/split.sh
@@ -2,8 +2,11 @@
 
 cd ../pgn
 
-../bin/pgn-extract -Wuci --nocomments --nonags --novars --nobadresults "$1" | \
-  ../bin/analyse --engine ../bin/stockfish --searchdepth 10 --annotatePGN | \
-  ../bin/pgn-extract -Wsan --addhashcode --totalplycount -#1,100000
+# ../bin/pgn-extract -Wuci --nocomments --nonags --novars --nobadresults "$1" | \
+#   ../bin/analyse --engine ../bin/stockfish --searchdepth 10 --annotatePGN | \
+#   ../bin/pgn-extract -Wsan --addhashcode --totalplycount -#1,100000
 
-rm "$1"
+
+../bin/pgn-extract -Wsan -#1,100000 \
+  --addhashcode --nocomments --nonags --novars --nobadresults \
+  "$1" && rm "$1"
